cast int

+++ imgui/binding/src/lua_imgui_enum.cpp
@@ -25,7 +25,7 @@ void imgui_binding_lua_register_enum(lua_State* L)
         for(auto& i : datas)
         {
             lua_pushlstring(L, i.name.data(), i.name.length());     // ? M k
-            lua_createtable(L, 0, i.data.size());                   // ? M k t
+            lua_createtable(L, 0, static_cast<int>(i.data.size())); // ? M k t
             for(auto& j : i.data)
             {
                 lua_pushlstring(L, j.name.data(), j.name.length()); // ? M k t k
